Completed exersize 05
this was decently simple, but i thought the read me asked to put more than one div around many others and i ignored that thinking it would be simpler with less, and i added two when i might've only needed one. i could've moved the close button into the header div, but instead i made a new one to surround the head and close button. the sizing isn't exact, but i dont think that's the point.